"Load" and "About" tab removal & potential viewer ID fix
- Cleaned up some comments in index.html
- Removed the "Load" and "About" tabs in viewers
- viewer.html to no longer references the glmol01_src textarea
- Modification of the naming algorithm for viewers in gridupdater.js
- May have fixed the naming problem?

#Known Bugs:
1. Chrome won't accept uploads to iframes when just running Hydra
locally.  Origins for the primary frame and iframe are both "null"
-> same origin violation.
2. Chrome and Firefox won't display more than 16-17 molecular
models even though the viewers themselves are still responsive.

#To Do:
- Show Webix ID of each viewer in the frame
- Streamline debugging and for user reference
- Fix above bug #1
- Allow grid downsizing
- Allow a file to be uploaded to a selected viewer

#Misc (from before):
If the mouse is being used for compound manipulation and is moved
outside the viewer before release, the viewer will continue interacting
with the mouse whenever cursor is in its bounds until the viewer is
clicked again.  Consider making a global mouse state variable in the
future?