Added grid downsizing + misc
- Changes made in gridupdater.js to allow grid to be shrunk
  -- Added logic checks: only relevant resizing will get called
- Commented out now extraneous code in viewer.html
- Attempted to add grid coordinate labels (eg "1,0") to viewers
  -- Could only get this to work with the initial viewer
  -- Added code not needed for labeling the initial viewer is commented
out

Known Bugs:
1) Chrome won't accept uploads to iframes when just running Hydra
locally.  Origins for the primary frame and iframe are both "null"
-> same origin violation.
2) Chrome and Firefox won't display more than 16-17 molecular
models even though the viewers themselves are still responsive.

To Do:
- Show Webix ID/grid coordinates in each viewer's frame
- Fix bug #1
- Allow a file to be uploaded to a selected viewer

#Misc (from before):
If the mouse is being used for compound manipulation and is moved
outside the viewer before release, the viewer will continue interacting
with the mouse whenever cursor is in its bounds until the viewer is
clicked again.  Consider making a global mouse state variable in the
future?
- Note: console now sometimes has the following error
-- "Error: WebGL: Exceeded 16 live WebGL contexts for this principal,
    losing the least recently used one."
-- Seems to be from changes made for displaying labels