Coordinate Display, Re-indexed Coordinates, & Misc Cleanup
- Fixed coordinate display
  -- Had to use a callback function in the onAfterLoad event
- Changed indexing of viewers so that coordinates start at 1
- Commented out the box that contained "Dock File" & "Ligand File"
- Removed now extraneous comments and commented out code

Known Bugs:
1) Chrome won't accept uploads to iframes when just running Hydra
locally.  Origins for the primary frame and iframe are both "null"
-> same origin violation.
-This actually works now? Possibly due to a Chrome update.
2) Chrome and Firefox won't display more than 16-17 molecular
models even though the viewers themselves are still responsive.
- "Error: WebGL: Exceeded 16 live WebGL contexts for this
principal, losing the least recently used one."
- This is due to limitations built into Three49custom.js

To Do:
- Improve controls for uploading files to viewers
-- Replace primitive controls with a file list + controls
-- Try implementing drag-and-drop
- Investigate mysterious resolving of bug #1
- Add padding between cells for easier scrolling on Macs
- Pull mouse mode setting (eg "Rotate", "Translate", etc)
out of viewers. Universal control in Hydra GUI
- Replace "ID" in "Compound List" with coordinates

Misc (from before):
If the mouse is being used for compound manipulation and is moved
outside the viewer before release, the viewer will continue interacting
with the mouse whenever cursor is in its bounds until the viewer is
clicked again.  Consider making a global mouse state variable in the
future?