Processor pulls Ligand ID/# + minor UI adjustments
The processor now pulls the ligand ID from the filename and uses that to
sort (ID shown in table). Does not rely on the filenames playing nice as
much.
Note that the ligand ID is NOT the ligand's ZINC ID.
Changed the way column widths get set in the processor window -> fill
space