Proto: Add props & vendors from DBs
Tentative code for grabbing property and vendor information from
database files in the codebase.
This code is called when an object in uploadTable (and therefore
comp_table too) is clicked for the first time:
  - Chemical properties + SMILES from 47_prop added as props
  - Vendors added as an object collection to the main object
  - If either chem props or vendors are not found, they are set to
    ' ' (space) so that the DB search code doesn't run on every click

The comp_prop pane is bound to uploadTable.
However, the vendor list gets a more manual clear -> new data

Discovered bug to fix:
Deleting a compound from uploadTable does not delete it from comp_table