Fix: "Delete Data" not removing from compound list
Removing an item from the uploadTable would previously only make the
corresponding cell in comp_table go blank.  This was because the other
table was just being refreshed.  This made the corresponding cell go
blank since it no longer had any data to reference, but it needed to
actually be told to do this after removing it from uploadTable